fix(ui): resolve services ingestion lazy split
