+++ openmetadata-ui/src/main/resources/ui/src/App.tsx
@@ -18,7 +18,6 @@ import { AuthProvider } from './components/Auth/AuthProviders/AuthProvider';
 import { queryClient } from './queryClient';
 
 const App: FC = () => {
-
   // QueryClientProvider sits ABOVE AuthProvider so that the singleton is available everywhere
   // — including AuthProvider's onLogout handler, which needs to clear the query cache so a
   // freshly-authenticated user can't see another principal's cached entity bodies. The
