+++ openmetadata-ui/src/main/resources/ui/src/components/Settings/Services/Ingestion/IngestionListTable/PipelineActions/PipelineActionsDropdown.tsx
@@ -28,7 +28,7 @@ import {
   IngestionPipeline,
   PipelineType,
 } from '../../../../../../generated/entity/services/ingestionPipelines/ingestionPipeline';
-import { getLoadingStatus } from '../../../../../../utils/EntityDisplayPureUtils';
+import { getLoadingStatus } from '../../../../../../utils/EntityDisplayUtils';
 import { getEntityName } from '../../../../../../utils/EntityNameUtils';
 import {
   getEditIngestionPath,
